requirements/base.txt
django==4.2.4
django-environ==0.10.0
psycopg2-binary==2.9.7
djangorestframework==3.14.0
drf-yasg==1.21.7
djangorestframework-simplejwt==5.2.2
django-import-export==3.3.1
requests==2.31.0
pandas==2.1.2